‘プロフィール画面のUI調整、ログアウトボタン設置’
